requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
paddlepaddle-gpu==2.6.2
paddleocr==2.7.3
numpy>=1.21.0
Pillow>=8.3.0
scikit-image>=0.18.0
shapely>=1.7.0
tqdm>=4.62.0
matplotlib>=3.3.0
# Additional dependencies that might be needed
scipy>=1.5.0
pandas>=1.1.0
# Removed problematic fsplit package